Deleted additional user class, Add classes in separate files.
