+++ User.cs
@@ -1,13 +1,16 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProjektBankenSquid2
+﻿namespace ProjektBankenSquid2
 {
-    internal class User
+    //Klass för användare
+    public class User
     {
-        public int ID { get; set; }
+        public string name { get; set; }
+        public int pincode { get; set; }
+
+        //Konstruktor
+        public User(string userName, int userPincode)
+        {
+            name = userName;
+            pincode = userPincode;
+        }
     }
 }
